Refactor evaluations: extract shared helpers and fix metric labels
- Move GroundTruthMissingError and has_ground_truth into _common.py to
  remove the circular-import workaround and the duplicated helper
- Filter phantom (support == 0) categories from per-category breakdown
- Rename score_dedup metric label from "f1" to "dedup_score" to match
  the actual formula (1 - normalized_delta)
- Document the 2-leg-only constraint in score_transfer_detection
