+++ src/moneybin/validation/evaluations/__init__.py
@@ -2,15 +2,9 @@
 
 from __future__ import annotations
 
-
-class GroundTruthMissingError(RuntimeError):
-    """Raised when an evaluation runs against a DB without `synthetic.ground_truth`."""
-
-
-from moneybin.validation.evaluations.categorization import (  # noqa: E402  # circular import — error class must be defined before submodule import
-    score_categorization,
-)
-from moneybin.validation.evaluations.matching import (  # noqa: E402  # circular import — error class must be defined before submodule import
+from moneybin.validation.evaluations._common import GroundTruthMissingError
+from moneybin.validation.evaluations.categorization import score_categorization
+from moneybin.validation.evaluations.matching import (
     score_dedup,
     score_transfer_detection,
 )
